some bug fixes

+++ src/gps/gps.h
@@ -14,13 +14,13 @@ public:
   bool update();
   bool getLocation(Location* out);
   bool getAltitude(double* alt);
-  double getTime();
+  uint32_t getTime();
   float getSatellites();
 
 private:
-  double baseTime;
-  double lastUpdated;
+  uint32_t baseTime;    // GPS time as Unix epoch seconds at the last valid fix.
+  uint32_t lastUpdated; // millis() captured at the last valid fix.
   TinyGPSPlus gps;
 
-  bool rawGetTime(double* out);
+  bool rawGetTime(uint32_t* out);
 };
